backend/ml-service/requirements.txt
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask==3.0.0
Flask-CORS==4.0.0
joblib==1.3.1
